requirement.txt
speechrecognition
PyAudio
RPi.GPIO
nltk
smbus2
pyjokes
request